Fix: Add since annotation in new functions and add WP version check in accordion inner blocks as well
