+++ plugins/woocommerce/src/Blocks/BlockTypes/Accordion/AccordionGroup.php
@@ -5,7 +5,6 @@ namespace Automattic\WooCommerce\Blocks\BlockTypes\Accordion;
 
 use Automattic\WooCommerce\Blocks\BlockTypes\AbstractBlock;
 use Automattic\WooCommerce\Blocks\BlockTypes\EnableBlockJsonAssetsTrait;
-use Automattic\WooCommerce\Blocks\Utils\Utils;
 
 /**
  * AccordionGroup class.
@@ -13,6 +12,7 @@ use Automattic\WooCommerce\Blocks\Utils\Utils;
 class AccordionGroup extends AbstractBlock {
 
 	use EnableBlockJsonAssetsTrait;
+	use AccordionVersionControlTrait;
 
 	/**
 	 * Block name.
@@ -26,39 +26,12 @@ class AccordionGroup extends AbstractBlock {
 	 *
 	 * - Hook into WP lifecycle.
 	 * - Register the block with WordPress.
+	 *
+	 * @since 10.4.3
 	 */
 	protected function initialize() {
-		add_filter( 'block_type_metadata_settings', array( $this, 'add_accordion_block_type_metadata_settings' ), 10, 2 );
+		$this->initialize_accordion_version_control();
 		parent::initialize();
-	}
-
-	/**
-	 * Modify block type metadata settings to hide accordion blocks from inserter on WordPress 6.9 or later.
-	 *
-	 * @param array $settings Array of determined settings for registering a block type.
-	 * @param array $metadata Metadata provided for registering a block type.
-	 *
-	 * @return array Modified settings.
-	 */
-	public function add_accordion_block_type_metadata_settings( $settings, $metadata ) {
-		// Check if this is the accordion group block.
-		if (
-			! empty( $metadata['name'] ) &&
-			'woocommerce/accordion-group' === $metadata['name']
-		) :
-			// Hide the accordion block from the inserter on WordPress 6.9 or later.
-			if ( Utils::wp_version_compare( '6.9', '>=' ) ) :
-				// Ensure supports array exists.
-				if ( ! isset( $settings['supports'] ) ) :
-					$settings['supports'] = array();
-				endif;
-
-				// Hide from inserter.
-				$settings['supports']['inserter'] = false;
-			endif;
-		endif;
-
-		return $settings;
 	}
 
 	/**
