+++ plugins/woocommerce/src/Blocks/BlockTypes/Accordion/AccordionItem.php
@@ -11,6 +11,7 @@ use Automattic\WooCommerce\Blocks\BlockTypes\EnableBlockJsonAssetsTrait;
 class AccordionItem extends AbstractBlock {
 
 	use EnableBlockJsonAssetsTrait;
+	use AccordionVersionControlTrait;
 
 	/**
 	 * Block name.
@@ -18,6 +19,19 @@ class AccordionItem extends AbstractBlock {
 	 * @var string
 	 */
 	protected $block_name = 'accordion-item';
+
+	/**
+	 * Initialize this block type.
+	 *
+	 * - Hook into WP lifecycle.
+	 * - Register the block with WordPress.
+	 *
+	 * @since 10.4.3
+	 */
+	protected function initialize() {
+		$this->initialize_accordion_version_control();
+		parent::initialize();
+	}
 
 	/**
 	 * Include and render the block.
