+++ plugins/woocommerce/src/Blocks/BlockTypes/Accordion/AccordionPanel.php
@@ -11,6 +11,7 @@ use Automattic\WooCommerce\Blocks\BlockTypes\EnableBlockJsonAssetsTrait;
 class AccordionPanel extends AbstractBlock {
 
 	use EnableBlockJsonAssetsTrait;
+	use AccordionVersionControlTrait;
 
 	/**
 	 * Block name.
@@ -18,4 +19,17 @@ class AccordionPanel extends AbstractBlock {
 	 * @var string
 	 */
 	protected $block_name = 'accordion-panel';
+
+	/**
+	 * Initialize this block type.
+	 *
+	 * - Hook into WP lifecycle.
+	 * - Register the block with WordPress.
+	 *
+	 * @since 10.4.3
+	 */
+	protected function initialize() {
+		$this->initialize_accordion_version_control();
+		parent::initialize();
+	}
 }
